[Impeller] Fix potential overflow when allocating buffers at the next power of two size (#188742)
Allocation::NextPowerOfTwoSize will now use 64-bit integers and will
check for inputs that are out of range.

See b/521829269

+++ engine/src/flutter/impeller/renderer/backend/vulkan/allocator_vk.cc
@@ -618,7 +618,7 @@ Bytes AllocatorVK::DebugGetHeapUsage() const {
     const VmaBudget& budget = budgets[i];
     total_usage += budget.usage;
   }
-  return Bytes{static_cast<double>(total_usage)};
+  return Bytes{total_usage};
 }
 
 void AllocatorVK::DebugTraceMemoryStatistics() const {
